+++ NEWS.md
@@ -1,3 +1,21 @@
+# RTL 0.1.4
+
+## New
++ `getCurve()` added to extract OHLC futures contract forward curves from `Morningstar`. 
++ `chart_eia_steo()` returns a Supply/Demand balance from the EIA STEO data set. Currently configured for Global Liquids and will be augmented for US Crude, Light and Middle Distillates.
++ `chart_eia_sd()` returns Supply/Demand balance from the EIA weekly data for mogas, distillates, jet and resids.
++ `tickers_eia` table updated to build Supply Demand Balances for US products.
++ Datasets `ref.opt.inputs` and `ref.opt.outputs` to support refinery LP optimization education using `lpSolve` package. 
++ `swapFutWeight()` returns the % applied to the first line contract in Calendar Month Average commodity swaps when two futures contracts are involved e.g. WTI. It uses the proper NYMEX or ICE holiday calendars and fit for purpose for building trading sheets.
++ `swapInfo()` returns all information required to price first line futures contract averaging swap or CMA physical trade, including a current month instrument with prior settlements.
+
+## Updates and Fixes
++ `eia2tidy()` fix for key variable in function.
++ `crudes` dataset updated.
+
+## Removed
+ + `twtrump` and `twoott` tweets datasets for learning NLP.
+
 # RTL 0.1.3
 
 + `eia2tidy()` removed dependency to `EIAdata` package.